scripts/requirements.txt
opencv-python
pillow
numpy
PyMuPDF
PyMuPDF>=1.24.0
Pillow>=10.0.0
reportlab